Address more comments
Create a new file to hold the TestData class. <-- This is a WIP.

Refs #9280
